requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
auth0-python==4.9.0
av==14.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.10
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
coqpit-config==0.2.0
coqui-tts==0.26.2
coqui-tts-trainer==0.2.3
crewai==0.121.0
cryptography==45.0.3
ctranslate2==4.6.0
curl_cffi==0.11.1
cycler==0.12.1
Cython==3.1.1
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
durationpy==0.10
einops==0.8.1
encodec==0.1.1
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.9
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
groq==0.25.0
grpcio==1.71.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
instructor==1.8.3
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
json5==0.12.0
json_repair==0.46.0
jsonlines==1.2.0
jsonpatch==1.33
jsonpickle==4.1.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
lazy_loader==0.4
librosa==0.11.0
litellm==1.68.0
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.41.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pocketsphinx==5.0.4
pooch==1.8.2
posthog==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrate-limiter==3.7.0
pysbd==0.3.4
pytest==8.3.5
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sec-edgar-api==1.1.0
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.45.3
streamlit==1.45.1
streamlit-webrtc==0.62.4
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.0
torchaudio==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
typeguard==4.4.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uv==0.7.8
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.61
zipp==3.22.0
zstandard==0.23.0
